requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
numpy==1.20.2
streamlit==0.62.0
Pillow==8.2.0
torch==1.8.1+cpu
torchvision==0.9.1+cpu
typing-extensions==3.10.0.0